@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -32,17 +32,26 @@ dependencies:
   http: ^0.13.1
   html: ^0.15.0
   carousel_slider: ^3.0.0
+  flutter_bloc: ^7.0.0
   custom_navigation_bar: ^0.8.1
+  bottom_navy_bar: ^6.0.0
+
   provider: ^5.0.0
-  date_range_picker: ^1.0.7
+  flutter_date_pickers: ^0.2.4
   firebase_auth: ^1.0.1
   cloud_firestore: ^1.0.3
   google_sign_in: ^5.0.1
   fluttertoast: ^8.0.3
   adaptive_dialog: ^0.10.0+5
-  onesignal_flutter: ^2.6.3
+  onesignal_flutter: ^3.0.0-beta2
+  mobx: ^2.0.1
+  flutter_mobx: ^2.0.0
 
 dev_dependencies:
+  build_runner: ^2.0.1
+  mobx_codegen: ^2.0.1+3
+  json_serializable: ^4.1.1
+  
   flutter_test:
     sdk: flutter
 
